enforce Hessian class for supporting dynamic graph
