+++ python/paddle/fluid/tests/unittests/autograd/test_autograd_functional.py
@@ -12,18 +12,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import collections
+import typing
 import unittest
-import paddle
-import numpy as np
 
-from paddle.autograd.functional import vjp, jvp, _as_tensors
-from paddle import grad, ones_like, zeros_like
-from funcs import reduce, reduce_dim, mul, pow, o2, unuse, nested, matmul
+import numpy as np
+import paddle
+from paddle.autograd.functional import _as_tensors, jvp, vjp
+
+import config
+import parameterize as param
+import utils
+from utils import matmul, mul, nested, o2, pow, reduce, reduce_dim, unuse
 
 
 def make_v(f, inputs):
     outputs = _as_tensors(f(*inputs))
-    return [ones_like(x) for x in outputs]
+    return [paddle.ones_like(x) for x in outputs]
 
 
 class TestAutogradFunctional(unittest.TestCase):
@@ -77,14 +82,14 @@ class TestAutogradFunctional(unittest.TestCase):
                 v = self.gen_inputs(v)
             outputs = func(*xs)
             if v is not None:
-                inputs_grad = grad(
+                inputs_grad = paddle.grad(
                     outputs,
                     xs,
                     v,
                     create_graph=create_graph,
                     allow_unused=allow_unused)
             else:
-                inputs_grad = grad(
+                inputs_grad = paddle.grad(
                     outputs,
                     xs,
                     create_graph=create_graph,
@@ -196,17 +201,17 @@ class TestVJP(TestAutogradFunctional):
 def jac(grad_fn, f, inputs):
     assert grad_fn in [vjp, jvp]
     if grad_fn is jvp:
-        vs = [zeros_like(x) for x in inputs]
+        vs = [paddle.zeros_like(x) for x in inputs]
     else:
         outputs = f(*inputs)
         if isinstance(outputs, paddle.Tensor):
             outputs = [outputs]
-        vs = [zeros_like(y) for y in outputs]
+        vs = [paddle.zeros_like(y) for y in outputs]
     JJ_cols = []
     for i, v in enumerate(vs):
         v = v.flatten()
         for j in range(len(v)):
-            _v = zeros_like(v).detach()
+            _v = paddle.zeros_like(v).detach()
             _v[j] = 1.0
             _v = _v.reshape(vs[i].shape)
             _vs = vs.copy()
@@ -260,9 +265,108 @@ class TestJVP(TestAutogradFunctional):
         for f, inputs in test_cases:
             inputs = self.gen_inputs(inputs)
             results_omitting_v = jvp(f, inputs)
-            v = [ones_like(x) for x in inputs]
+            v = [paddle.ones_like(x) for x in inputs]
             results_with_v = jvp(f, inputs, v)
             self.check_results(results_omitting_v, results_with_v)
+
+
+@param.place(config.DEVICES)
+@param.parameterize(
+    (param.TEST_CASE_NAME, 'func', 'xs'),
+    (  # noqa
+        ('1d-in-1d-out', utils.square, np.array([2., 3.])),
+        ('3d-in-3d-out', utils.square, np.random.rand(2, 3, 4)),
+        ('multi-input', utils.square, np.random.rand(10, 20)),
+        ('matmul', paddle.matmul,
+         (np.random.rand(2, 2), np.random.rand(2, 2))), ))
+class TestJacobianNoBatch(unittest.TestCase):
+    def setUp(self):
+        self._dtype = self.xs[0].dtype if isinstance(
+            self.xs, typing.Sequence) else self.xs.dtype
+        self._eps = config.EPS.get(str(self._dtype))
+
+        self.xs = [paddle.to_tensor(x) for x in self.xs] if isinstance(
+            self.xs, typing.Sequence) else paddle.to_tensor(self.xs)
+        self._actual = paddle.autograd.Jacobian(self.func, self.xs, False)
+        self._expected = self._expected()
+
+    def test_jacobian(self):
+        Index = collections.namedtuple('Index', ('type', 'value'))
+        indexes = (Index('all', (slice(0, None, None), slice(0, None, None))),
+                   Index('row', (0, slice(0, None, None))),
+                   Index('col', (slice(0, None, None), 0)),
+                   Index('multi-row', (slice(0, 2, 1), slice(0, None, None))))
+        self.assertEqual(self._actual[:].dtype, self._expected.dtype)
+        for index in indexes:
+            np.testing.assert_allclose(
+                self._actual.__getitem__(index.value),
+                self._expected.__getitem__(index.value),
+                rtol=config.RTOL.get(str(self._dtype)),
+                atol=config.ATOL.get(str(self._dtype)),
+                err_msg=f'Testcase {index.type} index not passed, value is {index.value}'
+            )
+
+    def _expected(self):
+        # numerical_jacobian return list of list of tensors, need to concat.
+        results = utils._compute_numerical_jacobian(self.func, self.xs,
+                                                    self._eps, self._dtype)
+        rows = []
+        for i in range(len(results)):
+            rows.append(
+                paddle.concat([paddle.to_tensor(x) for x in results[i]], -1))
+        return paddle.concat(rows, 0)
+
+
+@param.place(config.DEVICES)
+@param.parameterize(
+    (param.TEST_CASE_NAME, 'func', 'xs'),
+    (
+        ('1d-in-1d-out', utils.square, np.array([[1., 2., 3.], [3., 4., 3.]])),
+        ('3d-in-3d-out', utils.square, np.random.rand(2, 3, 4)),
+        ('multi-in-single-out', utils.square, np.random.rand(5, 6)),
+        # ('multi-in-multi-out', utils.o2, (np.random.rand(2,2), np.random.rand(2,2)))
+    ))
+class TestJacobianBatchFirst(unittest.TestCase):
+    def setUp(self):
+        self._dtype = self.xs[0].dtype if isinstance(
+            self.xs, typing.Sequence) else self.xs.dtype
+        self._eps = config.EPS.get(str(self._dtype))
+
+        self.xs = [paddle.to_tensor(x) for x in self.xs] if isinstance(
+            self.xs, typing.Sequence) else paddle.to_tensor(self.xs)
+        self._actual = paddle.autograd.Jacobian(self.func, self.xs, True)
+        self._expected = self._expected()
+
+    def test_jacobian(self):
+        Index = collections.namedtuple('Index', ('type', 'value'))
+        indexes = (
+            Index('all', (slice(0, None, None), slice(0, None, None),
+                          slice(0, None, None))),
+            Index('row', (slice(0, None, None), 0, slice(0, None, None))),
+            Index('col',
+                  (slice(0, None, None), slice(0, None, None), 0)), Index(
+                      'batch', (slice(0, 2, None), slice(0, None, None),
+                                slice(0, None, None))),
+            Index('multi-row',
+                  (slice(0, 1, None), slice(0, 2, 1), slice(0, None, None))))
+        self.assertEqual(self._actual[:].dtype, self._expected.dtype)
+        for index in indexes:
+            np.testing.assert_allclose(
+                self._actual.__getitem__(index.value),
+                self._expected.__getitem__(index.value),
+                rtol=config.RTOL.get(str(self._dtype)),
+                atol=config.ATOL.get(str(self._dtype)),
+                err_msg=f'Testcase {index.type} index not passed, value is {index.value}'
+            )
+
+    def _expected(self):
+        results = utils._compute_numerical_batch_jacobian(
+            self.func, self.xs, self._eps, self._dtype)
+        rows = []
+        for i in range(len(results)):
+            rows.append(
+                paddle.concat([paddle.to_tensor(x) for x in results[i]], -1))
+        return paddle.concat(rows, 1)
 
 
 if __name__ == "__main__":
